Update ending lecture 51. This includes the working jump mechanics.
Update ending lecture 51. This includes the working jump mechanics.

We should try to fix the jumping mechanics in the future. They are a bit choppy and need some work.  Mainly the underlying collider for the player isn't simulating physics properly. When it falls off object or goes over object it should bend a bit more and have some more give.